refactor(spec): adjust spec exports
