+++ packages/g6/src/index.ts
@@ -2,17 +2,4 @@ import './preset';
 
 export const version = '5.0.0';
 
-export type { G6Spec } from './spec';
-export type { BehaviorOptions } from './spec/behavior';
-export type { CanvasOptions } from './spec/canvas';
-export type { DataOptions } from './spec/data';
-export type { AnimationOptions } from './spec/element/animation';
-export type { ComboOptions } from './spec/element/combo';
-export type { EdgeOptions } from './spec/element/edge';
-export type { NodeOptions } from './spec/element/node';
-export type { PaletteOptions } from './spec/element/palette';
-export type { LayoutOptions } from './spec/layout';
-export type { OptimizeOptions } from './spec/optimize';
-export type { ThemeOptions } from './spec/theme';
-export type { ViewportOptions } from './spec/viewport';
-export type { WidgetOptions } from './spec/widget';
+export * from './spec';
